[WIP] yarn lint task WITH A TEMPORARY TRICK
NOTE that this has a trick to use eslint-config-standard-jsx
to quickly get a very basic working set of eslint
rules working with React that is up-to-date.

TODO: we need a new eslint config package that is designed to
gives us the basic set of working eslint rules
without using this kind of a trick.